Upload set-unit-age.lua
Adding set-unit-age.lua for setting unit age.  Best used on units that are transformed to different castes in the Guildhall or Garrison buildings in Masterwork DF.  This is because all transformations, even to other castes of the same species, cause the unit's age to be randomly set to somewhere around middle age for the species.  Thus a 12 year old dwarf getting guild training to join the smithing guild will suddenly end up 88 years old.